move AArch64 related code to AArch64.h
